requirements.txt
pafy==0.5.5
prompt-toolkit==1.0.14
pyfiglet==0.8.post1
Pygments==2.10.0
PyInquirer==1.0.3
python-dotenv==0.19.1
regex==2021.10.21
six==1.16.0
wcwidth==0.2.5
youtube-dl==2021.6.6
